agent/pyproject.toml
[tool.poetry]
name = "agent"
version = "0.5.2"
description = "The LLM Engine behind Tutor-GPT"
authors = ["Vineeth Voruganti <13438633+VVoruganti@users.noreply.github.com>"]
# readme = "README.md"
packages = [
  { include = "agent"},
]

[tool.poetry.dependencies]
python = "^3.11"
honcho-ai = "^0.0.14"
mirascope = "^0.18.0"
openai = "^1.3.8"
validators = "^0.20.0"
python-dotenv = "^1.0.1"
sentry-sdk = "^2.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
